Fix various compiler warnings

+++ jwz/remote.c
@@ -265,13 +265,13 @@ send_xscreensaver_command (Display *dpy, Atom command, long arg,
 		    }
 
 		if (any && nhacks == 1)
-		  fprintf (stdout, " (hack #%ld)\n", data[2]);
+		  fprintf (stdout, " (hack #%ld)\n", (long)data[2]);
 		else if (any)
 		  {
 		    fprintf (stdout, " (hacks: ");
 		    for (i = 0; i < nhacks; i++)
 		      {
-			fprintf (stdout, "#%ld", data[2 + i]);
+			fprintf (stdout, "#%ld", (long)data[2 + i]);
 			if (i != nhacks-1)
 			  fputs (", ", stdout);
 		      }
